Fix: Resolve FriendCard component error
The "Component is not a function" error has been resolved by ensuring `FriendCard` is correctly exported and memoized. Import paths have been verified, and `FriendCard` has been made backward compatible to support both click and non-click usage patterns. Error boundary protection has been added for better debugging.
